Tweaks to sequence boundaries
Round start of seq to a gulp size.
Decide if a sequence has ended by checking if the first packet of the next sequence
is the current time step + gulp size

This enforces gulp_size = slot_size. But, it's not obvious slot size actually
does anything other than set the time granularity at which a buffer may start