Implement source blanking.
This is actually quite annoying. The call to blank a source
doesn't know about the details of the packet structure, and one can't
necessarily blank the right parts of the receive buffer when only given
nsrc, src, nchan, chan, since these values have been abused. E.g.
the output buffer is in Time x Chan x Antpol order, but the Chan
dimension can be a multiple of nchan, and nsrc != the number of Antpols in the system.

Hack the information out of the packet stream and then write a clurgy blanking
function.

Barely tested.