Configure as a nix flake
It provides an overlay for nixpkgs that adds ctypesgen and a
configurable bifrost package, that can be overridden with various
versions of python3 and cudatoolkit (or without them), and can enable
debugging or not.  The github workflow uses nix to build a few
variations, run non-GPU tests, and update documentation in gh-pages
branch. Builds are cached by cachix; upon merge that part may need
some keys loaded onto the ledatelescope repo. Maybe needs a blurb in
the README before next release.