Resolve class/struct distinction warning from clang
For these classes, we're declaring them a typedef with `struct NAME *`
in a `.h` file so that it can be used from C. But then the body is
defined in C++ with `class`. The clang compiler on Mac complains (but
says "I'm okay with this, but Microsoft might not be"). So it doesn't
really matter, but I'd rather get rid of the warnings if it doesn't
cost anything.

It's okay to use `struct` in C++ with inheritance and all the other
features of classes... the main difference is that `struct` members
are public by default vs `class` being private by default.

So here we turn all of these that use the typedef struct pointer C
into structs rather than classes, which silences the warnings.

```
In file included from ring.cpp:34:
./ring_impl.hpp:47:1: warning: class 'BFspan_impl' was previously declared as a
      struct; this is valid, but may result in linker errors under the Microsoft
      C++ ABI [-Wmismatched-tags]
class BFspan_impl;
^
./bifrost/ring.h:67:16: note: previous use is here
typedef struct BFspan_impl*        BFspan;
               ^
```