My contribution to reducing the build burden?
It may be neat that we can ensure it works with different releases of
nixpkgs, and survey the horizon of nix-unstable. But any errors that
pop up there (but not in pinned) are probably not our issue. Compiling
against carefully selected variations of python and C++ might be more
productive.