Compiler-dependent bug fix for Socket get_family
Both the Linux and BSD man pages for getsockname(2) say:

   "The addrlen argument should be initialized to indicate the amount of
    space (in bytes) pointed to by addr."
    https://man7.org/linux/man-pages/man2/getsockname.2.html
    https://developer.apple.com/library/archive/documentation/System/Conceptual/ManPages_iPhoneOS/man2/getsockname.2.html

We weren't doing that, so len was uninitialized and you get what you
get. On some versions of some compilers we got away with it. We seemed
to get away with it more reliably on Linux than on Mac. But when it goes
wrong, then get_family doesn't return the right thing. Subsequently,
get_mtu can't find the appropriate interface to report on, so it ends up
returning 0 and that caused test_address to fail. Booyah!