Resolve potential IPV6 memcmp issue
Again, Mac-only code... but gcc warned:

```
Socket.cpp: In function ‘int get_mtu(int)’:
Socket.cpp:82:24: warning: ‘int
   __builtin_memcmp_eq(const void*, const void*, long unsigned int)’
      reading 16 bytes from a region of size 8 [-Wstringop-overflow=]
         82 |         if( std::memcmp(inaddr6->sin6_addr.s6_addr,
	 addr6->sin6_addr.s6_addr, 16) == 0  )
```

So I wondered, why would it be that addr6's s6_addr is only size 8? Well
addr6 is a reinterpret cast of addr which is a plain sockaddr struct
declared locally. I take it the way to ensure there's enough space is to
use sockaddr_storage here, and that fixed the warning.